drivers: power: reset: Fix adb reboot bootloader/recovery
This change is based on OnePlus3 kernel which made
a hack to force a soft reset as this preserves
memory contents, including pstore and other
persistent memory.

In this case we are focing the soft reset only if the bootloader or
recovery is sent as cmd.

Reference: https://github.com/LineageOS/android_kernel_oneplus_msm8996/commit/f0979313e42697b3bc2311f0241cfb045c18eed0

Signed-off-by: Humberto Borba <humberos@gmail.com>
Change-Id: Ibf201de9cdceab8434926d9e35e453c526c14056
Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>