usb: udc: ci13xxx: Fix cable disconnection IRQ panic
On MSM platform we have experienced a bad interrupt handler
system error because the controller wasn't getting suspended
properly, hence we had an interrupt continuously firing.

Move the suspend interrupt handling in the udc_irq function
and update the "suspended" status variable early to avoid
any missed suspend in external suspend handlers and any
spurious interrupts due to workers trying to send data to
USB when the controller is suspended.