f2fs rapidly gc all invalid blocks when screen goes off & plugged in
Wait until the screen goes off and detect if the power source is plugged in.
If it is, GC thread runs repeatedly with 1ms interval.
Since the device can go to sleep even with power source plugged in,
catch an additional wakelock (ew, a wakelock).
This patch is intended to optimize background GC behavior specific to Android.
So, disable background_gc by default as well.

Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>