defconfig: kanuti: Enable stackprotector strong
Enable protection against stack corruption. (most) Functions with the
possibility of stack corruption  are protected with a canary.

With this change, the Kanuti kernel grows from 7064639 to 7212880
an increase of 2.1%. Performance loss is considered minimal. Security
against stack overflow is greatly improved. [1]

This is the improved version of CONFIG_CC_STACKPROTECTOR=y which is
enabled in Kanuti's kernel.

[1] https://lwn.net/Articles/584225

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>