arm: DT: msm8939-camera: Add bus vectors in JPEG nodes
Update JPEG DT entries to k3.18 camera for JPEG to work on k3.10.

The bus vectors are statically setup in jpeg driver which is not
portable on non-msm platforms. Hence add the vectors in DT node where
the vectors can be setup with the bus routine msm_bus_cl_get_pdata()
dynamically.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>