arm: DT: Cleanup and reorganization for msm8939 tulip
Since caf does not support msm8939 on 3.18 kernel, there is no
point in supporting msm8939-mtp, so let's clean this up by removing
the unused hardware nodes, and moving the ones we use to the
parent device trees.

The sdhc2 gpio hack is left in the tulip device tree because it needs
to be fixed properly.