arm: DT: msm8939-common: Add tsens client-id
Use this flag to identify the SW ID mapping used to associate it with
the controller and the physical sensor mapping within the controller.
The physical sensor mapping within each controller is done using the
qcom,sensor-id property. If the property is not present the SW ID
mapping with default from 0 to total number of supported sensors with
each controller instance.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>