regulator: onsemi-ncp6335d: Change initcall level for early probe
ACPU clock driver initialization has dependency on regulator
driver initialization.

The cpufreq driver to successfully register requires the acpu
clock handle which in turn requires the apc regulator handle.

In a recent change the init-call ordering of the cpufreq and
acpu-clock drivers has changed to register them early. Move the
onsemi regulator initialization to arch_initcall() to handle
this modification.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>