msm: kgsl: Use fault context to retrieve process information
Instead of looking up the process by pagetable base and
name, use the fault context to extract the pid and other
process specific information. This works for both the
per-process and global pagetable configurations and also
reduces some locking. This also reports the correct
pid and task name in the global pagetable configuration.msm: kgsl_legacy: Refactor MMU/IOMMU support

The current MMU code assumes a binary state - either there is a
IOMMU or there isn't. This precludes other memory models and
makes for a lot of inherent IOMMU knowledge in the generic MMU
code and the rest of the driver.  Reorganize and cleanup the
MMU and IOMMU code:

 - Add a Kconfig boolean dependent on  ARM and/or MSM SMMU support.
 - Make "nommu" mode an actual MMU subtype and figure out available
   MMU subtypes at probe time.
 - Move IOMMU device tree parsing to the IOMMU code.
 - Move the MMU subtype private structures into struct kgsl_mmu.
 - Move adreno_iommu specific functions out of other generic
   adreno code.
 - Move A4XX specific preemption code out of the ringbuffer code.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>