msm: kgsl: Implement legacy KGSL driver
This legacy kgsl driver comes from LA.HB.1.3.2 branch and
works great with old blobs made for 3.10 kernel.

The new 3.18+ KGSL kernel driver uses a new way to manage
Adreno ringbuffer, which is implemented in userspace
libraries from the new UM branches.

This driver is the only way to get MSM8939 to work on 3.18+
kernels, due to the new one requiring a new ringbuffer
implementation, which requires new Adreno GPU firmwares,
which is not available for old A4xx chips.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>