msm: kgsl_legacy: Do not free gpuaddr range if unmap fails
Currently, we free up the gpuaddr range whether unmap succeeds
or fails. If it fails and we free up the gpuaddr range, there
is a possibility that a later allocation will overlap with this
range. Mapping this new range leads to a BUG_ON() in the iommu
driver which thinks that we are overwriting an existing mapping.
To avoid this, we never free the failing gpuaddr range, effectively
removing it from being re-used for any future mappings.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>