msm: kgsl_legacy: Add array of page pointers to memdesc
This is done to improve the kgsl vmfault routine. Currently,
it traverses the sglist to find the faulted page, which takes
linear time. By having an array of all the page pointers,
this operation will be completed in constant time.
Also, allocate sgt only for mapping this memory to the GPU.

Since this optimization is not needed for secure/global or
imported memory, we will not keep this array but keep
the sgt instead.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>