clk: msm: Fix for cpufreq for MSM8939 clock driver
After adding support for single cluster SoC in k3.18 we
can't initialize CPU on MSM8939.

The PLL reconfiguration is different for the single cluster SoC.
We have already successfully reconfigured MUX_CCI here:
if (!single_cluster) {
	rate = clk_get_rate(&cci_clk.c);
	clk_set_rate(&cci_clk.c, rate);
}

So we don't need to do this again on the non single cluster SoC.
In our case on MSM8939.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>