arm: DT: msm8936/39: Add support for partial goods CPU
Add support to thermal driver to handle partial goods.

In this case, it is possible that some physical CPUs
may not be brought online and it can be decided runtime.
This behavior changes the physical CPU to logical CPU
numbering dynamic.

Thermal driver should be capable to map the temperature
sensor with the correct logical CPU number. This is
achieved by creating different mitigation profiles and
associating each physical CPU to one profile using
device tree phandle. Thus thermal driver identifies the
thermal profile associated with the logical CPU and then
determines the mitigation information. The mitigation
profile has a phandle to the temperature sensor and
devicetree properties to enable or disable various type
of mitigation. A single device can have multiple sensors
mapped in array of phandles or even multiple sensors mapping
to a single sensor.

Partial goods cpu feature will be enabled only if
'qcom,mitigation-profile-enable' DT node is defined
in msm_thermal node.

Signed-off-by: Pavel Dubrova <pashadubrova@gmail.com>